modified .gitignore and got repo ready with running server on port 5001 and connection with neon db
