+++ .env
@@ -1,1 +1,2 @@
+PORT=5001
 DATABASE_URL=postgresql://PokemonBattleGame_owner:npg_L2eljRF7YsMg@ep-damp-forest-a2e5li68-pooler.eu-central-1.aws.neon.tech/PokemonBattleGame?sslmode=require
